Add a Jepsen test of the lock servertt
