+++ atlasdb-jepsen-tests/src/jepsen/atlasdb/timestamp.clj
@@ -1,54 +1,16 @@
-(ns jepsen.atlasdb
-  (:require [clojure.tools.logging :refer :all]
+(ns jepsen.atlasdb.timestamp
+  (:require [jepsen.atlasdb.timelock :as timelock]
             [jepsen.checker :as checker]
             [jepsen.client :as client]
-            [jepsen.control :as c]
-            [jepsen.db :as db]
             [jepsen.generator :as gen]
             [jepsen.nemesis :as nemesis]
             [jepsen.os.debian :as debian]
+            [jepsen.tests :as tests]
             [jepsen.util :refer [timeout]]
-            [knossos.history :as history]
-            [jepsen.tests :as tests])
+            [knossos.history :as history])
     ;; We can import any Java objects, since Clojure runs on the JVM
     (:import com.palantir.atlasdb.jepsen.JepsenHistoryCheckers)
     (:import com.palantir.atlasdb.http.TimestampClient))
-
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-;; Server setup, teardown, and log files
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-(defn create-server
-  "Creates an object that implements the db/DB protocol.
-   This object defines how to setup and teardown a timelock server on a given node, and specifies where the log files
-   can be found.
-  "
-  []
-  (reify db/DB
-    (setup! [_ _ node]
-      (c/su
-        (debian/install-jdk8!)
-        (info node "Uploading and unpacking timelock server")
-        (c/upload "resources/atlasdb/atlasdb-timelock-server.tgz" "/")
-        (c/exec :mkdir "/atlasdb-timelock-server")
-        (c/exec :tar :xf "/atlasdb-timelock-server.tgz" "-C" "/atlasdb-timelock-server" "--strip-components" "1")
-        (c/upload "resources/atlasdb/timelock.yml" "/atlasdb-timelock-server/var/conf")
-        (c/exec :sed :-i (format "s/<HOSTNAME>/%s/" (name node)) "/atlasdb-timelock-server/var/conf/timelock.yml")
-        (info node "Starting timelock server")
-        (c/exec :env "JAVA_HOME=/usr/lib/jvm/java-8-oracle" "/atlasdb-timelock-server/service/bin/init.sh" "start")
-        (info node "Waiting until timelock cluster is ready")
-        (TimestampClient/waitUntilHostReady (name node))
-        (Thread/sleep (* 1000 10))
-        (TimestampClient/waitUntilTimestampClusterReady '("n1" "n2" "n3" "n4" "n5"))))
-
-    (teardown! [_ _ node]
-      (c/su
-        (try (c/exec "/atlasdb-timelock-server/service/bin/init.sh" "stop") (catch Exception _))
-        (try (c/exec :rm :-rf "/atlasdb-timelock-server") (catch Exception _))
-        (try (c/exec :rm :-f "/atlasdb-timelock-server.tgz") (catch Exception _))))
-
-    db/LogFiles
-    (log-files [_ test node]
-      ["/atlasdb-timelock-server/var/log/atlasdb-timelock-server-startup.log"])))
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;; Defining the set of of operations that you can do with a client
@@ -95,7 +57,7 @@
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;; Defining the Jepsen test
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-(defn atlasdb-test
+(defn timestamp-test
   []
   (assoc tests/noop-test
     :os debian/os
@@ -109,5 +71,5 @@
                                      (gen/sleep 85)
                                      {:type :info, :f :stop}])))
                     (gen/time-limit 360))
-    :db (create-server)
+    :db (timelock/create-db)
     :checker checker))
